d08-django-3-advanced/requirements.txt
django
django-bootstrap5
psycopg2-binary
python-dateutil
